Add the opus 4.5 prototype
Prompt:

You are going to use beads to manage todos. Run `bd quickstart` before continuing on to the instructinos below.

Using https://ragelang.com/blog/2025/12/15/introduction-to-ragelang/  as the source, build out a tested, typescript interpreter for the language described. It should also have canvas support for outputting the sprites and text as described.